Consolidated versioning logic in core::version
Moved `storage::version` to `core::version` to unify versioning primitives
(NodeVersion, EdgeVersion, etc.) with `VersionMetadata`. This breaks the
artificial split between version metadata and data, strengthening the
core domain definition.

- Merged `src/storage/version.rs` into `src/core/version.rs`
- Deleted `src/storage/version.rs`
- Updated `src/storage/mod.rs` to re-export `core::version`
- Updated global imports from `crate::storage::version` to `crate::core::version`
- Preserved `storage::version` path for backward compatibility via re-export
